in progress

+++ Backend/Migrations/BarterContextModelSnapshot.cs
@@ -129,8 +129,8 @@ namespace barterserv.Migrations
                         .IsRequired()
                         .HasColumnType("longtext CHARACTER SET utf8mb4");
 
-                    b.Property<int>("Rating")
-                        .HasColumnType("int");
+                    b.Property<float>("Rating")
+                        .HasColumnType("float");
 
                     b.Property<int>("ReceiverId")
                         .HasColumnType("int");
@@ -155,7 +155,7 @@ namespace barterserv.Migrations
                             AuthorId = 7,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Très satisfait du service rendu,je recommande",
-                            Rating = 4,
+                            Rating = 4f,
                             ReceiverId = 8,
                             ServiceLinkedToId = 3
                         },
@@ -165,7 +165,7 @@ namespace barterserv.Migrations
                             AuthorId = 7,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Prestation excellente ,vraiment au dessus de nos attentes.Je recommande à 100%",
-                            Rating = 5,
+                            Rating = 5f,
                             ReceiverId = 8,
                             ServiceLinkedToId = 3
                         },
@@ -175,7 +175,7 @@ namespace barterserv.Migrations
                             AuthorId = 3,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Médiocre",
-                            Rating = 1,
+                            Rating = 1f,
                             ReceiverId = 8,
                             ServiceLinkedToId = 1
                         },
@@ -185,7 +185,7 @@ namespace barterserv.Migrations
                             AuthorId = 2,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "peu recommandable",
-                            Rating = 1,
+                            Rating = 1f,
                             ReceiverId = 8,
                             ServiceLinkedToId = 1
                         },
@@ -195,7 +195,7 @@ namespace barterserv.Migrations
                             AuthorId = 4,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Reussit",
-                            Rating = 3,
+                            Rating = 3f,
                             ReceiverId = 8,
                             ServiceLinkedToId = 2
                         },
@@ -205,7 +205,7 @@ namespace barterserv.Migrations
                             AuthorId = 4,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "peu mieux faire",
-                            Rating = 3,
+                            Rating = 3f,
                             ReceiverId = 8,
                             ServiceLinkedToId = 4
                         },
@@ -215,7 +215,7 @@ namespace barterserv.Migrations
                             AuthorId = 5,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Recomandable",
-                            Rating = 4,
+                            Rating = 4f,
                             ReceiverId = 8,
                             ServiceLinkedToId = 5
                         },
@@ -225,7 +225,7 @@ namespace barterserv.Migrations
                             AuthorId = 6,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Prestation excellente ",
-                            Rating = 5,
+                            Rating = 5f,
                             ReceiverId = 8,
                             ServiceLinkedToId = 6
                         },
@@ -235,7 +235,7 @@ namespace barterserv.Migrations
                             AuthorId = 7,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "je recommande",
-                            Rating = 4,
+                            Rating = 4f,
                             ReceiverId = 8,
                             ServiceLinkedToId = 7
                         },
@@ -245,7 +245,7 @@ namespace barterserv.Migrations
                             AuthorId = 9,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "vraiment au dessus de nos attentes",
-                            Rating = 5,
+                            Rating = 5f,
                             ReceiverId = 1,
                             ServiceLinkedToId = 8
                         },
@@ -255,7 +255,7 @@ namespace barterserv.Migrations
                             AuthorId = 9,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Très satisfait",
-                            Rating = 5,
+                            Rating = 5f,
                             ReceiverId = 1,
                             ServiceLinkedToId = 9
                         },
@@ -265,7 +265,7 @@ namespace barterserv.Migrations
                             AuthorId = 1,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Je recommande à 100%",
-                            Rating = 5,
+                            Rating = 5f,
                             ReceiverId = 2,
                             ServiceLinkedToId = 10
                         },
@@ -275,7 +275,7 @@ namespace barterserv.Migrations
                             AuthorId = 3,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "service rendu excellent",
-                            Rating = 4,
+                            Rating = 4f,
                             ReceiverId = 2,
                             ServiceLinkedToId = 11
                         },
@@ -285,7 +285,7 @@ namespace barterserv.Migrations
                             AuthorId = 4,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Prestation à recommande à 100%",
-                            Rating = 5,
+                            Rating = 5f,
                             ReceiverId = 2,
                             ServiceLinkedToId = 12
                         },
@@ -295,7 +295,7 @@ namespace barterserv.Migrations
                             AuthorId = 5,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Très bien fait",
-                            Rating = 4,
+                            Rating = 4f,
                             ReceiverId = 3,
                             ServiceLinkedToId = 13
                         },
@@ -305,7 +305,7 @@ namespace barterserv.Migrations
                             AuthorId = 6,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = " Nos attente non pas étaient déçues",
-                            Rating = 5,
+                            Rating = 5f,
                             ReceiverId = 3,
                             ServiceLinkedToId = 14
                         },
@@ -315,7 +315,7 @@ namespace barterserv.Migrations
                             AuthorId = 7,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "A recommande",
-                            Rating = 4,
+                            Rating = 4f,
                             ReceiverId = 3,
                             ServiceLinkedToId = 15
                         },
@@ -325,7 +325,7 @@ namespace barterserv.Migrations
                             AuthorId = 8,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Prestation moyenne",
-                            Rating = 3,
+                            Rating = 3f,
                             ReceiverId = 4,
                             ServiceLinkedToId = 16
                         },
@@ -335,7 +335,7 @@ namespace barterserv.Migrations
                             AuthorId = 9,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Très mitigé",
-                            Rating = 2,
+                            Rating = 2f,
                             ReceiverId = 4,
                             ServiceLinkedToId = 17
                         },
@@ -345,7 +345,7 @@ namespace barterserv.Migrations
                             AuthorId = 2,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Trés amateur hélas!!",
-                            Rating = 1,
+                            Rating = 1f,
                             ReceiverId = 4,
                             ServiceLinkedToId = 18
                         },
@@ -355,7 +355,7 @@ namespace barterserv.Migrations
                             AuthorId = 6,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Prester apres 2 rdv ,mais bien executé",
-                            Rating = 3,
+                            Rating = 3f,
                             ReceiverId = 5,
                             ServiceLinkedToId = 19
                         },
@@ -365,7 +365,7 @@ namespace barterserv.Migrations
                             AuthorId = 7,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "probleme de retard à l heure convenue ,mais prestation super",
-                            Rating = 4,
+                            Rating = 4f,
                             ReceiverId = 5,
                             ServiceLinkedToId = 20
                         },
@@ -375,7 +375,7 @@ namespace barterserv.Migrations
                             AuthorId = 1,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = " Satisfait ",
-                            Rating = 3,
+                            Rating = 3f,
                             ReceiverId = 5,
                             ServiceLinkedToId = 21
                         },
@@ -385,7 +385,7 @@ namespace barterserv.Migrations
                             AuthorId = 8,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Rien de pas ordinnaire",
-                            Rating = 3,
+                            Rating = 3f,
                             ReceiverId = 6,
                             ServiceLinkedToId = 22
                         },
@@ -395,7 +395,7 @@ namespace barterserv.Migrations
                             AuthorId = 9,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Service minimum",
-                            Rating = 2,
+                            Rating = 2f,
                             ReceiverId = 6,
                             ServiceLinkedToId = 23
                         },
@@ -405,7 +405,7 @@ namespace barterserv.Migrations
                             AuthorId = 5,
                             Date = new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified),
                             Description = "Passable",
-                            Rating = 3,
+                            Rating = 3f,
                             ReceiverId = 6,
                             ServiceLinkedToId = 24
                         });
